small fix5

+++ src/App/FractalWindow.h
@@ -31,7 +31,7 @@ protected:
 	void wheelEvent(QWheelEvent * e) override;
 
 private:
-	const float wheelSensitivity = 1.06;
+	const float wheelSensitivity = static_cast<float>(1.06);
 
 	GLint iterUniform_ = -1;
 	GLint bailOutUniform_ = -1;
